+++ test/index.js
@@ -20,6 +20,8 @@ var sequelizeMySQL = new Sequelize(
 
 var PieStatGetter = require('../services/pie-stat-getter');
 var LineStatGetter = require('../services/line-stat-getter');
+var ResourcesGetter = require('../services/resources-getter');
+var HasManyGetter = require('../services/has-many-getter');
 
 [sequelizePostgres, sequelizeMySQL].forEach(function (sequelize) {
   var models = {};
@@ -39,6 +41,17 @@ var LineStatGetter = require('../services/line-stat-getter');
     resetPasswordToken: { type: Sequelize.STRING }
   });
 
+  models.address = sequelize.define('address', {
+    line: { type: Sequelize.STRING },
+    zipCode: { type: Sequelize.STRING },
+    city: { type: Sequelize.STRING },
+    country: { type: Sequelize.STRING },
+    userId: { type: Sequelize.INTEGER }
+  });
+
+  models.address.belongsTo(models.user);
+  models.user.hasMany(models.address);
+
   Interface.Schemas = {
     schemas: {
       user: {
@@ -54,7 +67,24 @@ var LineStatGetter = require('../services/line-stat-getter');
           { field: 'password', type: 'String' },
           { field: 'createdAt', type: 'Date' },
           { field: 'updatedAt', type: 'Date' },
-          { field: 'resetPasswordToken', type: 'String' }
+          { field: 'resetPasswordToken', type: 'String' },
+          { field: 'addresses', type: ['Number'] }
+        ]
+      },
+      address: {
+        name: 'address',
+        idField: 'id',
+        primaryKeys: ['id'],
+        isCompositePrimary: false,
+        fields: [
+          { field: 'id', type: 'Number' },
+          { field: 'line', type: 'String' },
+          { field: 'zipCode', type: 'String' },
+          { field: 'city', type: 'String' },
+          { field: 'country', type: 'String' },
+          { field: 'user', type: 'Number', references: 'user.id' },
+          { field: 'createdAt', type: 'Date' },
+          { field: 'updatedAt', type: 'Date' }
         ]
       }
     }
@@ -70,7 +100,7 @@ var LineStatGetter = require('../services/line-stat-getter');
       });
 
       describe('A simple Pie Chart on an empty users table', function () {
-        it('should repond with an empty value', function (done) {
+        it('should generate a valid SQL query', function (done) {
           return new PieStatGetter(models.user, {
               type: 'Pie',
               collection: 'user',
@@ -91,7 +121,7 @@ var LineStatGetter = require('../services/line-stat-getter');
       });
 
       describe('A simple Line Chart per day on an empty users table', function () {
-        it('should repond with an empty value', function (done) {
+        it('should generate a valid SQL query', function (done) {
           return new LineStatGetter(models.user, {
               type: 'Line',
               collection: 'user',
@@ -110,9 +140,11 @@ var LineStatGetter = require('../services/line-stat-getter');
             });
         });
       });
+    });
 
+    describe('Stats > Line Stat Getter', function () {
       describe('A simple Line Chart per week on an empty users table', function () {
-        it('should repond with an empty value', function (done) {
+        it('should generate a valid SQL query', function (done) {
           return new LineStatGetter(models.user, {
               type: 'Line',
               collection: 'user',
@@ -133,7 +165,7 @@ var LineStatGetter = require('../services/line-stat-getter');
       });
 
       describe('A simple Line Chart per month on an empty users table', function () {
-        it('should repond with an empty value', function (done) {
+        it('should generate a valid SQL query', function (done) {
           return new LineStatGetter(models.user, {
               type: 'Line',
               collection: 'user',
@@ -154,7 +186,7 @@ var LineStatGetter = require('../services/line-stat-getter');
       });
 
       describe('A simple Line Chart per year on an empty users table', function () {
-        it('should repond with an empty value', function (done) {
+        it('should generate a valid SQL query', function (done) {
           return new LineStatGetter(models.user, {
               type: 'Line',
               collection: 'user',
@@ -171,6 +203,210 @@ var LineStatGetter = require('../services/line-stat-getter');
               expect(stat.value.length).equal(0);
               done();
             });
+        });
+      });
+    });
+
+    describe('Resources > Resources Getter', function () {
+      describe('Request on the resources getter without page size', function () {
+        it('should generate a valid SQL query', function (done) {
+          var params = {
+            fields: {
+              user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken'
+            },
+            page: { number: '1' },
+            timezone: '+02:00'
+          };
+          return new ResourcesGetter(models.user, { sequelize: sequelize }, params)
+            .perform()
+            .then(function () {
+              done();
+            });
+        });
+      });
+
+      describe('Request on the resources getter with a page size', function () {
+        it('should generate a valid SQL query', function (done) {
+          var params = {
+            fields: {
+              user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken'
+            },
+            page: { number: '1', size: '30' },
+            timezone: '+02:00'
+          };
+          return new ResourcesGetter(models.user, { sequelize: sequelize }, params)
+            .perform()
+            .then(function () {
+              done();
+            });
+        });
+      });
+
+      describe('Request on the resources getter with a sort on the primary key', function () {
+        it('should generate a valid SQL query', function (done) {
+          var params = {
+            fields: {
+              user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken'
+            },
+            sort: '-id',
+            page: { number: '1', size: '30' },
+            timezone: '+02:00'
+          };
+          return new ResourcesGetter(models.user, { sequelize: sequelize }, params)
+            .perform()
+            .then(function () {
+              done();
+            });
+        });
+      });
+
+      describe('Request on the resources getter with a search', function () {
+        it('should generate a valid SQL query', function (done) {
+          var params = {
+            fields: {
+              user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken'
+            },
+            page: { number: '1', size: '30' },
+            search: 'hello',
+            timezone: '+02:00'
+          };
+          return new ResourcesGetter(models.user, { sequelize: sequelize }, params)
+            .perform()
+            .then(function () {
+              done();
+            });
+        });
+      });
+
+      describe('Request on the resources getter with a "not contains" filter condition', function () {
+        it('should generate a valid SQL query', function (done) {
+          var params = {
+            fields: {
+              user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken'
+            },
+            page: { number: '1', size: '30' },
+            filterType: 'and',
+            filter: { username: '!*hello*' },
+            timezone: '+02:00'
+          };
+          return new ResourcesGetter(models.user, { sequelize: sequelize }, params)
+            .perform()
+            .then(function () {
+              done();
+            });
+        });
+      });
+
+      describe('Request on the resources getter with a filter condition and search', function () {
+        it('should generate a valid SQL query', function (done) {
+          var params = {
+            fields: {
+              user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken'
+            },
+            page: { number: '2', size: '50' },
+            filterType: 'and',
+            filter: { username: '*hello*' },
+            search: 'world',
+            timezone: '+02:00'
+          };
+          return new ResourcesGetter(models.user, { sequelize: sequelize }, params)
+            .perform()
+            .then(function () {
+              done();
+            });
+        });
+      });
+
+      describe('Request on the resources getter with a filter condition, search and sort combined', function () {
+        it('should generate a valid SQL query', function (done) {
+          var params = {
+            fields: {
+              user: 'id,firstName,lastName,username,password,createdAt,updatedAt,resetPasswordToken'
+            },
+            page: { number: '2', size: '50' },
+            filterType: 'and',
+            filter: { username: '*hello*' },
+            sort: '-id',
+            search: 'world',
+            timezone: '+02:00'
+          };
+          return new ResourcesGetter(models.user, { sequelize: sequelize }, params)
+            .perform()
+            .then(function () {
+              done();
+            });
+        });
+      });
+
+      describe('HasMany > HasMany Getter', function () {
+        before(function (done) {
+          return models.user
+            .create({ email: 'louis@forestadmin.com' })
+            .then(function () {
+              return done();
+            });
+        });
+
+        describe('Request on the hasMany getter without sort', function () {
+          it('should generate a valid SQL query', function (done) {
+            var params = {
+              recordId: 1,
+              associationName: 'addresses',
+              fields: {
+                address: 'line,zipCode,city,country,user'
+              },
+              page: { number: '1', size: '20' },
+              timezone: '+02:00'
+            };
+            return new HasManyGetter(models.user, models.address, { sequelize: sequelize }, params)
+              .perform()
+              .then(function (result) {
+                expect(result[0]).equal(0);
+                done();
+              });
+          });
+        });
+
+        describe('Request on the hasMany getter with a sort on an attribute', function () {
+          it('should generate a valid SQL query', function (done) {
+            var params = {
+              recordId: 1,
+              associationName: 'addresses',
+              fields: {
+                address: 'line,zipCode,city,country,user'
+              },
+              page: { number: '1', size: '20' },
+              sort: 'city',
+              timezone: '+02:00'
+            };
+            return new HasManyGetter(models.user, models.address, { sequelize: sequelize }, params)
+              .perform()
+              .then(function (result) {
+                expect(result[0]).equal(0);
+                done();
+              });
+          });
+        });
+
+        describe('Request on the hasMany getter with a sort on a belongsTo', function () {
+          it('should generate a valid SQL query', function (done) {
+            var params = {
+              recordId: 1,
+              associationName: 'addresses',
+              fields: {
+                address: 'line,zipCode,city,country,user'
+              },
+              page: { number: '1', size: '20' },
+              sort: '-user.id',
+              timezone: '+02:00'
+            };
+            return new HasManyGetter(models.user, models.address, { sequelize: sequelize }, params)
+              .perform()
+              .then(function (result) {
+                expect(result[0]).equal(0);
+                done();
+              });
+          });
         });
       });
     });